feat: Add validity checking functions (#12)
Really we want to keep validity checking on pretty much always for
unknown input, since the behaviour of invalid geographies is
unpredictable in terms of containment or overlay. On the flip side, GEOS
doesn't really validate anything (although sometimes has unpredictable
results when these get put into some functions). In any case, when these
errors do come up, it's helpful to be able to diagnose them.

`s2_geogfromtext/wkb_novalidate()` is a bit long...this could eventually
be a named parameter like it is in R (`validate = false`).